Small fix to make the lines pixel-perfect in chrome
When looking at the pattern in Chrome 49, you'll notice that the small
sides are less wide than the other lines. You won't see this in IE and
this change does nothing for IE (must be a rounding issue), however,
using this updated style will make sure Chrome renders it the same as IE
(and just looks better).